WebScraper/requirements.txt
psycopg2
sqlalchemy
attrs
nameparser
lxml
selenium
